refactor: refactor directory structure for BadgeList and FilterButton
