+++ src/features/foods/components/index.ts
@@ -1,7 +1,5 @@
 export { AddButton } from './AddButton';
 export { BadgeList } from './BadgeList';
-export { CategoryBadge } from './CategoryBadge';
-export { FilterBadge } from './FilterBadge';
 export { FilterButton } from './FilterButton';
 export { FoodList } from './FoodList';
 export { FoodSort } from './FoodSort';
